backend/pyproject.toml
[project]
name = "bedrock_test_backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "cryptography>=45.0.3",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.11",
    "sphinx>=8.2.3",
]
